backend-flask/requirements.txt
flask
flask-cors
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
# AWS cloudwatch
watchtower
# AWS X-Ray
aws-xray-sdk
# Rollbar
rollbar
blinker